refactor: public 폴더 구조화
