+++ bubbletea03-my-app/src/component/home/Footer.js
@@ -2,11 +2,11 @@ function Footer() {
     return (
         <div className="footer">
             <div className="home-button">
-                <img src="home_icon.png"/>
+                <img src="home/home_icon.png"/>
                 <h2>홈</h2>
             </div>
             <div className="mycarrot-button">
-                <img src="person_icon.png"/>
+                <img src="home/person_icon.png"/>
                 <h2>나의 당근</h2>
             </div>
         </div>
