+++ bubbletea03-my-app/src/component/home/Header.js
@@ -1,0 +1,10 @@
+function Header() {
+    return (
+        <div className="header">
+            <h2>홈</h2>
+            <img src="home/hamburger_icon.png"/>
+        </div>
+    );
+}
+
+export default Header;
