+++ bubbletea03-my-app/src/component/home/PlusButton.js
@@ -1,7 +1,7 @@
 function PlusButton() {
     return (
         <div className="plus-button">
-            <img src="plus_icon.png"/>
+            <img src="home/plus_icon.png"/>
         </div>
     );
 }
