+++ bubbletea03-my-app/src/component/product-page/Header.js
@@ -1,7 +1,7 @@
 function Header() {
     return (
         <div className="header">
-            <h2>홈</h2>
+            <h2></h2>
             <img src="hamburger_icon.png"/>
         </div>
     );
