+++ bubbletea03-my-app/src/page/ProductPage.js
@@ -1,0 +1,10 @@
+function ProductPage() {
+    return (
+        <div className="product-page">
+            <h2>홈</h2>
+            <img src="hamburger_icon.png"/>
+        </div>
+    );
+}
+
+export default ProductPage;
